peer-manager: extract event helpers
